Moved Peg related utils to utils package to defeat out-of-package instantiation
