+++ rskj-core/src/main/java/co/rsk/peg/BridgeStorageProvider.java
@@ -25,6 +25,7 @@ import co.rsk.core.RskAddress;
 import co.rsk.crypto.Keccak256;
 import co.rsk.peg.bitcoin.CoinbaseInformation;
 import co.rsk.peg.fastbridge.FastBridgeFederationInformation;
+import co.rsk.peg.utils.BridgeSerializationUtils;
 import co.rsk.peg.whitelist.LockWhitelist;
 import co.rsk.peg.whitelist.LockWhitelistEntry;
 import co.rsk.peg.whitelist.OneOffWhiteListEntry;
