+++ rskj-core/src/main/java/co/rsk/peg/utils/PegUtils.java
@@ -1,8 +1,5 @@
 package co.rsk.peg.utils;
 
-import co.rsk.peg.BridgeSerializationUtils;
-import co.rsk.peg.BridgeUtils;
-
 public class PegUtils {
 
     private final BridgeUtils bridgeUtils;
