+++ rskj-core/src/main/java/co/rsk/remasc/RemascContract.java
@@ -21,7 +21,7 @@ package co.rsk.remasc;
 import co.rsk.config.RemascConfig;
 import co.rsk.core.RskAddress;
 import co.rsk.panic.PanicProcessor;
-import co.rsk.peg.BridgeSerializationUtils;
+import co.rsk.peg.utils.BridgeSerializationUtils;
 import co.rsk.rpc.modules.trace.ProgramSubtrace;
 import org.ethereum.config.Constants;
 import org.ethereum.config.blockchain.upgrades.ActivationConfig;
