+++ rskj-core/src/test/java/co/rsk/peg/performance/RegisterBtcCoinbaseTransactionTest.java
@@ -23,7 +23,7 @@ import co.rsk.bitcoinj.script.Script;
 import co.rsk.bitcoinj.store.BlockStoreException;
 import co.rsk.bitcoinj.store.BtcBlockStore;
 import co.rsk.peg.BridgeMethods;
-import co.rsk.peg.BridgeSerializationUtils;
+import co.rsk.peg.utils.BridgeSerializationUtils;
 import co.rsk.peg.BridgeStorageProvider;
 import co.rsk.peg.utils.PegUtils;
 import org.bouncycastle.util.encoders.Hex;
