+++ rskj-core/src/test/java/co/rsk/remasc/RemascRskAddressActivationTest.java
@@ -21,7 +21,7 @@ package co.rsk.remasc;
 import co.rsk.config.RemascConfig;
 import co.rsk.config.RemascConfigFactory;
 import co.rsk.core.RskAddress;
-import co.rsk.peg.BridgeSerializationUtils;
+import co.rsk.peg.utils.BridgeSerializationUtils;
 import co.rsk.peg.utils.PegUtils;
 import org.ethereum.config.Constants;
 import org.ethereum.config.blockchain.upgrades.ActivationConfig;
